+++ test/fixture/mock-generator/final-method/builder.php
@@ -1,0 +1,7 @@
+<?php
+
+return new Eloquent\Phony\Mock\Builder\MockBuilder(
+    'Eloquent\Phony\Test\TestClassF',
+    null,
+    'Phony\Test\MockGeneratorFinalMethod'
+);